Extract handlers to files
