+++ src/ForEvolve.ExceptionFilters.Abstraction/BaseHttpExceptions/Handlers/InternalServerErrorExceptionHandler.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.AspNetCore.Http;
+
+namespace ForEvolve.ExceptionFilters.Handlers
+{
+    public class InternalServerErrorExceptionHandler : ExceptionHandler<InternalServerErrorException>
+    {
+        public override int StatusCode => StatusCodes.Status500InternalServerError;
+    }
+}
