+++ src/ForEvolve.ExceptionFilters.Abstraction/BaseHttpExceptions/Handlers/NotFoundExceptionHandler.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.AspNetCore.Http;
+
+namespace ForEvolve.ExceptionFilters.Handlers
+{
+    public class NotFoundExceptionHandler : ExceptionHandler<NotFoundException>
+    {
+        public override int StatusCode => StatusCodes.Status404NotFound;
+    }
+}
